+++ server/service/globals.php
@@ -1,0 +1,17 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+
+<?php
+    define('FITROCKR_UPDATE_USER', 'fitrockr_update_user');
+    /* Page keywords */
+    define('SH_MODULE_FITROCKR', 'sh_module_fitrockr');
+
+    /* API URLs */
+    define('FITROCKR_URL_CREATE_USER', 'https://api.fitrockr.com/v1/users');
+
+    /* Table names */
+    define('TABLE_USERS_FITROCKR', 'users_fitrockr');
+?>