+++ CHANGELOG.md
@@ -6,3 +6,5 @@
  - add page `sh_global_fitrockr`
   - field `fitrockr_api_key`
   - field `fitrockr_api_tenant`
+ - add update `Fitrockr` user
+ - automatically create `Fitrockr` user on validate and link it to Selfhelp